requirements.txt
tk==0.1.0
marshmallow==3.15.0
opencv-python==4.5.5.64
opencv-contrib-python==4.5.5.64
pydash==5.1.0
Pillow==9.1.0
requests==2.27.1